backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
motor==3.3.2
pymongo==4.6.0
openai==1.3.7
azure-storage-blob==12.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
pillow>=10.1.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
PyMuPDF>=1.23.8

